Added sprinting functionality
When the player holds shift, or if you want to set it to a controller
just also assign the run binding to a controller button. When the player
runs he drains stamina, when it hits 0 there is a 3 second cooldown
where he runs super slow then it regenerates again.

I also added UMG for the stamina hud.